수정/삭제 보안처리 기능 추가

+++ src/main/java/net/slipp/web/QuestionController.java
@@ -52,15 +52,34 @@ public class QuestionController {
 	}
 	
 	@GetMapping("/{id}/form")
-	public String updateform(@PathVariable Long id, Model model) {
-		model.addAttribute("question", questionRepository.findById(id).get());
+	public String updateform(@PathVariable Long id, Model model, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepository.findById(id).get();
+		if(!question.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
+		
+		model.addAttribute("question", question);
 		
 		return "qna/updateForm";
 	}
 	
 	@PostMapping("/{id}")
-	public String update(@PathVariable Long id, String title, String contents) {
+	public String update(@PathVariable Long id, String title, String contents, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
 		Question question = questionRepository.findById(id).get();
+		if(!question.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
+		
 		question.update(title, contents);
 		questionRepository.save(question);
 		
@@ -68,7 +87,17 @@ public class QuestionController {
 	}
 	
 	@PostMapping("/{id}/delete")
-	public String delete(@PathVariable Long id, String title) {
+	public String delete(@PathVariable Long id, String title, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) {
+			return "/users/loginForm";
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepository.findById(id).get();
+		if(!question.isSameWriter(loginUser)) {
+			return "/users/loginForm";
+		}
+		
 		questionRepository.deleteById(id);
 
 		return "redirect:/";
